Fixing adset delete and update logic, push ads to zones when accepted

+++ controllers/adSet.js
@@ -79,10 +79,10 @@ async function updateById(req, res) {
         .status(config.status_code.FORBIDEN)
         .send({ message: "wrong user" });
     }
-    let { result, adOfferDoc } = adSetSupport.getAdSetStatus(id);
-    if (result == false) {
+    let { result, adOfferDoc } = await adSetSupport.getAdSetStatus(id);
+    if (result === false) {
       return res.status(config.status_code.FORBIDEN).send({
-        message: "adSet is using by some adOffer or playlist in adOffer",
+        message: "Some active ads are using this setting",
         adOffers: adOfferDoc,
       });
     }
@@ -101,11 +101,18 @@ async function updateById(req, res) {
 async function deleteById(req, res) {
   try {
     const { id } = req.params;
-    let document = adSetService.getById(id);
+    let document = await adSetService.getById(id);
     if (document["adManagerId"].toString() != req.userId) {
       return res
         .status(config.status_code.FORBIDEN)
         .send({ message: "wrong user" });
+    }
+    let { result, adOfferDoc } = await adSetSupport.getAdSetStatus(id);
+    if (result === false) {
+      return res.status(config.status_code.FORBIDEN).send({
+        message: "Some active ads are using this setting",
+        adOffers: adOfferDoc,
+      });
     }
     await adSetService.deleteById(id);
     return res.status(config.status_code.OK).send({ adset: true });
